start splitting autoprecompiles
